requirements.txt
torch==1.10.1
torchaudio==0.10.1
librosa
matplotlib==3.4.3
numpy==1.21.2
omegaconf==2.1.1
pandas==1.3.2
SoundFile==0.10.3.post1
speechbrain==0.5.9
hydra-core
jiwer
tqdm